Adjusted CNN Model (excluding as it is too large), corrected confidence metrics

+++ backend/backend/src/main/java/com/backend/backend/lesson/dto/WireframeResponse.java
@@ -1,3 +1,3 @@
 package com.backend.backend.lesson.dto;
 
-public record WireframeResponse(String prediction, Integer confidence, String wireframe) {}
+public record WireframeResponse(String prediction, Double confidence, String wireframe) {}
